Pending changes exported from your codespace

+++ src/main/java/org/vmallocator/AllocationInfo.java
@@ -5,9 +5,9 @@ import java.math.BigDecimal;
 public record AllocationInfo(
                 int vmSize,
                 int numberOfVMs,
-                int allocatableCPUsPerVM,
-                int requestedCPUsPerVM,
-                int totalIdleCPUs,
+                double allocatableCPUsPerVM,
+                double requestedCPUsPerVM,
+                double totalIdleCPUs,
                 int totalBillableCPUs,
                 BigDecimal totalCostPerHour) {
 }
